timeout.c created: move client timeouts code there.
